+++ aiohttp_things/__init__.py
@@ -9,7 +9,7 @@ from aiohttp_things.views import (
     PrimaryKeyMixin,
 )
 
-__version__ = '0.8.0'
+__version__ = '0.9.0'
 __all__ = [
     'views',
     'ContextMixin',